transfer category view
