+++ adminpage/sport/static/sport/js/modules/networking.mjs
@@ -51,5 +51,9 @@ async function sendResults(url, data, method = 'POST') {
 }
 
 function goto_profile() {
+    window.location.href = "/django/profile";
+}
+
+function reload_page() {
     window.location.reload();
 }
