+++ adminpage/sport/urls.py
@@ -10,4 +10,5 @@ urlpatterns = [
 
     path('profile/', profile_view, name='profile'),
     path('category/', category_view, name='category'),
+    path('calendar/sport/<int:sport_id>', calendar_view, name="sport_schedule_calendar"),
 ]
